feat: add min dist between spawning objects through grid (with jitter) and min dist from actor when spawning

+++ doom_creator/util/make.py
@@ -154,7 +154,7 @@ def make_acs(
     metabolic_delay: int,
     metabolic_damage: int,
     spawn_relative: bool = False,
-    spawn_range: float = 1000.0,
+    spawn_range: int = 1000,
 ):
     """Creates the acs script determining spawning and behaviour of all actors"""
     object_variables_acs = ""
